move install/uninstall functions to utils
