+++ utils.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
+	"math"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -12,6 +13,8 @@ import (
 	"time"
 
 	"github.com/levigross/grequests"
+	"github.com/dchest/uniuri"
+	"github.com/pbnjay/memory"
 
 	_ "github.com/lib/pq"
 )
@@ -51,6 +54,74 @@ func check(e error) {
 	if e != nil {
 		log.Fatal(e)
 	}
+}
+
+func uninstall() {
+	check(runCmd("docker", "stack", "rm", "utmstack"))
+
+	// sleep while docker is removing the containers
+	time.Sleep(120 * time.Second)
+
+	// remove images
+	for _, image := range containersImages {
+		check(runCmd("docker", "rmi", "utmstack.azurecr.io/"+image))
+	}
+
+	// logout from registry
+	runCmd("docker", "logout", "utmstack.azurecr.io")
+}
+
+func installProbe(datadir, pass, host string) {
+	logstashPipeline := mkdirs(0777, datadir, "logstash", "pipeline")
+	logsDir := mkdirs(0777, datadir, "logs")
+
+	serverName, err := os.Hostname()
+	check(err)
+	env := []string{
+		"SERVER_TYPE=probe",
+		"SERVER_NAME=" + serverName,
+		"DB_HOST=" + host,
+		"DB_PASS=" + pass,
+		"LOGSTASH_PIPELINE=" + logstashPipeline,
+		"UTMSTACK_LOGSDIR=" + logsDir,
+	}
+
+	initDocker(baseTemplate, env)
+}
+
+func installMaster(datadir, pass, fqdn, customerName, customerEmail string) {
+	esData := mkdirs(0777, datadir, "elasticsearch", "data")
+	esBackups := mkdirs(0777, datadir, "elasticsearch", "backups")
+	nginxCert := mkdirs(0777, datadir, "nginx", "cert")
+	logstashPipeline := mkdirs(0777, datadir, "logstash", "pipeline")
+	logsDir := mkdirs(0777, datadir, "logs")
+
+	serverName, err := os.Hostname()
+	check(err)
+	secret := uniuri.NewLenChars(10, []byte("abcdefghijklmnopqrstuvwxyz0123456789"))
+	env := []string{
+		"SERVER_TYPE=aio",
+		"SERVER_NAME=" + serverName,
+		"DB_PASS=" + pass,
+		"CLIENT_SECRET=" + secret,
+		fmt.Sprint("ES_MEM=", (memory.TotalMemory()/uint64(math.Pow(1024, 3))-4)/2),
+		"ES_DATA=" + esData,
+		"ES_BACKUPS=" + esBackups,
+		"NGINX_CERT=" + nginxCert,
+		"LOGSTASH_PIPELINE=" + logstashPipeline,
+		"UTMSTACK_LOGSDIR=" + logsDir,
+	}
+
+	initDocker(masterTemplate, env)
+
+	// Generate nginx auto-signed cert and key
+	generateCerts(nginxCert)
+
+	// configure elastic
+	initializeElastic(secret)
+
+	// Initialize PostgreSQL Database
+	initializePostgres(pass, customerName, fqdn, secret, customerEmail)
 }
 
 func initDocker(composerTemplate string, env []string) {
